fix msvc warning

+++ src/platform/sdl2.c
@@ -796,8 +796,8 @@ void platform_poll_events(Client *c) {
         } break;
 #endif
         case SDL_FINGERMOTION: {
-            float x = e.tfinger.x * SCREEN_FB_WIDTH;
-            float y = e.tfinger.y * SCREEN_FB_HEIGHT;
+            int x = (int)e.tfinger.x * SCREEN_FB_WIDTH;
+            int y = (int)e.tfinger.y * SCREEN_FB_HEIGHT;
 
             c->shell->idle_cycles = 0;
             c->shell->mouse_x = x;
@@ -808,8 +808,8 @@ void platform_poll_events(Client *c) {
             }
         } break;
         case SDL_FINGERDOWN: {
-            float x = e.tfinger.x * SCREEN_FB_WIDTH;
-            float y = e.tfinger.y * SCREEN_FB_HEIGHT;
+            int x = (int)e.tfinger.x * SCREEN_FB_WIDTH;
+            int y = (int)e.tfinger.y * SCREEN_FB_HEIGHT;
 
             c->shell->idle_cycles = 0;
             // set mouse pos here again due to no mouse movement
